+++ epam-promo/components/pickers/docs/pickerInput.doc.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { PickerTogglerProps, PickerInputBaseProps } from '@epam/uui-components';
 import { DocBuilder, isReadonlyDoc } from '@epam/uui-docs';
 import { PickerInput, PickerInputProps } from '../PickerInput';
 import { Button, LinkButton } from '../../buttons';
@@ -6,13 +7,12 @@ import { SearchInput } from '../../inputs';
 import { iEditable, isDisabledDoc } from '../../../docs';
 import { DefaultContext, ResizableContext, GridContext, FormContext } from '../../../docs';
 import { pickerBaseOptionsDoc } from './common';
-import { PickerTogglerProps, PickerInputBaseProps } from '@epam/uui-components';
-import {FlexCell} from "../../layout/FlexItems";
-import {Text} from "../../typography";
+import { FlexCell } from '../../layout/FlexItems';
+import { Text } from '../../typography';
 
 const PickerInputDoc = new DocBuilder<PickerInputBaseProps<any, any> & PickerInputProps>({ name: 'PickerInput', component: PickerInput })
     .implements([isDisabledDoc, isReadonlyDoc, iEditable, pickerBaseOptionsDoc] as any)
-    .prop('size', { examples: ['24', '30', '36', '42'], defaultValue: '36' })
+    .prop('size', { examples: ['24', '30', '36', '42', '48'], defaultValue: '36' })
     .prop('value', { examples: [
             { name: '1', value: 1 },
             { name: '[1, 2]', value: [1, 2] },